LU-11838 ldlm: struct timespec64.tv_sec type change
Since kernel 4.18 struct timespec64 is no longer defined
as struct timespec on 64bit systems, this means tv_sec
is no longer __kernel_time_t but now time64_t.

Use %llu as the format specifier and explicitly cast it
to unsigned long long.

Test-Parameters:trivial
Reviewed-on: https://review.whamcloud.com/34677
Tested-by: Jenkins
Tested-by: Maloo <maloo@whamcloud.com>

+++ lustre/ldlm/ldlm_lockd.c
@@ -2431,10 +2431,10 @@ static int ldlm_cancel_handler(struct ptlrpc_request *req)
 	if (req->rq_export == NULL) {
 		struct ldlm_request *dlm_req;
 
-		CERROR("%s from %s arrived at %lu with bad export cookie %llu\n",
+		CERROR("%s from %s arrived at %llu with bad export cookie %llu\n",
 		       ll_opcode2str(lustre_msg_get_opc(req->rq_reqmsg)),
 		       libcfs_nid2str(req->rq_peer.nid),
-		       req->rq_arrival_time.tv_sec,
+		       (unsigned long long)req->rq_arrival_time.tv_sec,
 		       lustre_msg_get_handle(req->rq_reqmsg)->cookie);
 
 		if (lustre_msg_get_opc(req->rq_reqmsg) == LDLM_CANCEL) {
